Host any leading content
If we have some code before the the first aware/props, it's probably important for adjusting state before those things evaluate. An example of this is the Avatar component.

Instead of having to rethink how all of that works, this change simply hoists that content so the component does not need to be reworked in order to use "optimized"